+++ src/agents/sandbox/tool-policy.test.ts
@@ -1,14 +1,13 @@
 import { describe, expect, it } from "vitest";
-import type { OpenClawConfig } from "../config/config.js";
-import { resolveSandboxConfigForAgent } from "./sandbox/config.js";
-import { resolveSandboxRuntimeStatus } from "./sandbox/runtime-status.js";
+import type { OpenClawConfig } from "../../config/config.js";
+import { resolveSandboxConfigForAgent } from "./config.js";
 import {
-  formatEffectiveSandboxToolPolicyBlockedMessage,
-  isToolAllowedBySandboxToolPolicy,
-  resolveEffectiveSandboxToolPolicyForAgent,
-} from "./tool-policy-sandbox.js";
+  formatSandboxToolPolicyBlockedMessage,
+  resolveSandboxRuntimeStatus,
+} from "./runtime-status.js";
+import { isToolAllowed, resolveSandboxToolPolicyForAgent } from "./tool-policy.js";
 
-describe("tool-policy-sandbox", () => {
+describe("sandbox/tool-policy", () => {
   it("merges sandbox alsoAllow into the default sandbox allowlist", () => {
     const cfg: OpenClawConfig = {
       agents: {
@@ -30,7 +29,7 @@ describe("tool-policy-sandbox", () => {
       },
     };
 
-    const resolved = resolveEffectiveSandboxToolPolicyForAgent(cfg, "tavern");
+    const resolved = resolveSandboxToolPolicyForAgent(cfg, "tavern");
     expect(resolved.allow).toContain("message");
     expect(resolved.allow).toContain("tts");
     expect(resolved.sources.allow).toEqual({
@@ -55,14 +54,17 @@ describe("tool-policy-sandbox", () => {
       },
     };
 
-    const resolved = resolveEffectiveSandboxToolPolicyForAgent(cfg, "main");
+    const resolved = resolveSandboxToolPolicyForAgent(cfg, "main");
     expect(resolved.allow).toContain("browser");
     expect(resolved.deny).not.toContain("browser");
     expect(
-      isToolAllowedBySandboxToolPolicy("browser", {
-        allow: resolved.allow,
-        deny: resolved.deny,
-      }),
+      isToolAllowed(
+        {
+          allow: resolved.allow,
+          deny: resolved.deny,
+        },
+        "browser",
+      ),
     ).toBe(true);
   });
 
@@ -83,20 +85,26 @@ describe("tool-policy-sandbox", () => {
       },
     };
 
-    const resolved = resolveEffectiveSandboxToolPolicyForAgent(cfg, "main");
+    const resolved = resolveSandboxToolPolicyForAgent(cfg, "main");
     expect(resolved.allow).toEqual([]);
     expect(resolved.deny).not.toContain("browser");
     expect(
-      isToolAllowedBySandboxToolPolicy("read", {
-        allow: resolved.allow,
-        deny: resolved.deny,
-      }),
+      isToolAllowed(
+        {
+          allow: resolved.allow,
+          deny: resolved.deny,
+        },
+        "read",
+      ),
     ).toBe(true);
     expect(
-      isToolAllowedBySandboxToolPolicy("browser", {
-        allow: resolved.allow,
-        deny: resolved.deny,
-      }),
+      isToolAllowed(
+        {
+          allow: resolved.allow,
+          deny: resolved.deny,
+        },
+        "browser",
+      ),
     ).toBe(true);
   });
 
@@ -158,20 +166,26 @@ describe("tool-policy-sandbox", () => {
       },
     };
 
-    const resolved = resolveEffectiveSandboxToolPolicyForAgent(cfg, "main");
+    const resolved = resolveSandboxToolPolicyForAgent(cfg, "main");
     expect(resolved.deny).toContain("browser");
     expect(resolved.deny).toContain("message");
     expect(
-      isToolAllowedBySandboxToolPolicy("browser", {
-        allow: resolved.allow,
-        deny: resolved.deny,
-      }),
+      isToolAllowed(
+        {
+          allow: resolved.allow,
+          deny: resolved.deny,
+        },
+        "browser",
+      ),
     ).toBe(false);
     expect(
-      isToolAllowedBySandboxToolPolicy("message", {
-        allow: resolved.allow,
-        deny: resolved.deny,
-      }),
+      isToolAllowed(
+        {
+          allow: resolved.allow,
+          deny: resolved.deny,
+        },
+        "message",
+      ),
     ).toBe(false);
   });
 
@@ -191,7 +205,7 @@ describe("tool-policy-sandbox", () => {
       },
     };
 
-    const browserMessage = formatEffectiveSandboxToolPolicyBlockedMessage({
+    const browserMessage = formatSandboxToolPolicyBlockedMessage({
       cfg,
       sessionKey: "agent:main:main",
       toolName: "browser",
@@ -199,11 +213,74 @@ describe("tool-policy-sandbox", () => {
     expect(browserMessage).toContain('Tool "browser" blocked by sandbox tool policy');
     expect(browserMessage).toContain("tools.sandbox.tools.deny");
 
-    const messageToolMessage = formatEffectiveSandboxToolPolicyBlockedMessage({
+    const messageToolMessage = formatSandboxToolPolicyBlockedMessage({
       cfg,
       sessionKey: "agent:main:main",
       toolName: "message",
     });
     expect(messageToolMessage).toBeUndefined();
+  });
+
+  it("keeps blocked-tool guidance glob-aware and shell-safe", () => {
+    const sessionKey = "agent:main:weird session;rm -rf /";
+    const cfg: OpenClawConfig = {
+      agents: {
+        defaults: {
+          sandbox: { mode: "all", scope: "agent" },
+        },
+      },
+      tools: {
+        sandbox: {
+          tools: {
+            deny: ["WEB_*"],
+          },
+        },
+      },
+    };
+
+    const message = formatSandboxToolPolicyBlockedMessage({
+      cfg,
+      sessionKey,
+      toolName: "web_fetch",
+    });
+
+    expect(message).toContain('Tool "web_fetch" blocked by sandbox tool policy');
+    expect(message).toContain("tools.sandbox.tools.deny");
+    expect(message).not.toContain(`Session: ${sessionKey}`);
+    expect(message).toContain("Session: agent:… -rf /");
+    expect(message).toContain(
+      "openclaw sandbox explain --session 'agent:main:weird session;rm -rf /'",
+    );
+  });
+
+  it("avoids terminal injection for control-character session keys", () => {
+    const sessionKey = "agent:main:abcde\n12345";
+    const cfg: OpenClawConfig = {
+      agents: {
+        defaults: {
+          sandbox: { mode: "all", scope: "agent" },
+        },
+      },
+      tools: {
+        sandbox: {
+          tools: {
+            deny: ["browser"],
+          },
+        },
+      },
+    };
+
+    const message = formatSandboxToolPolicyBlockedMessage({
+      cfg,
+      sessionKey,
+      toolName: "browser",
+    });
+
+    const sessionLine = message?.split("\n").find((line) => line.startsWith("Session: "));
+    expect(sessionLine).toBeDefined();
+    expect(sessionLine).not.toContain(sessionKey);
+    expect(sessionLine).toContain("\\n");
+    expect(message).toContain("openclaw sandbox explain --agent main");
+    expect(message).not.toContain("--session");
   });
 });
